# shared/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Al Qattan Curtains showroom API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: categoryId
          in: query
          required: false
          schema: { type: integer }
        - name: search
          in: query
          required: false
          schema: { type: string }
        - name: featured
          in: query
          required: false
          schema: { type: boolean }
        - name: page
          in: query
          required: false
          schema: { type: integer }
        - name: limit
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductListResponse" }
    post:
      operationId: createProduct
      tags: [products, admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
  /products/{id}:
    patch:
      operationId: updateProduct
      tags: [products, admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
    delete:
      operationId: deleteProduct
      tags: [products, admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /products/by-slug/{slug}:
    get:
      operationId: getProductBySlug
      tags: [products]
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /products/{id}/related:
    get:
      operationId: getRelatedProducts
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Related products
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Product" }
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UploadUrlRequest" }
      responses:
        "200":
          description: Upload URL
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UploadUrlResponse" }
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    Product:
      type: object
      required: [id, name, nameEn, slug, description, specs, price, priceText, categoryId, categoryName, images, isFeatured, isAvailable, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        nameEn: { type: ["string", "null"] }
        slug: { type: string }
        description: { type: ["string", "null"] }
        specs: { type: ["string", "null"] }
        price: { type: ["number", "null"] }
        priceText: { type: ["string", "null"] }
        categoryId: { type: ["integer", "null"] }
        categoryName: { type: ["string", "null"] }
        images:
          type: array
          items: { type: string }
        isFeatured: { type: boolean }
        isAvailable: { type: boolean }
        createdAt: { type: string, format: date-time }
    ProductInput:
      type: object
      required: [name, slug]
      properties:
        name: { type: string }
        nameEn: { type: ["string", "null"] }
        slug: { type: string }
        description: { type: ["string", "null"] }
        specs: { type: ["string", "null"] }
        price: { type: ["number", "null"] }
        priceText: { type: ["string", "null"] }
        categoryId: { type: ["integer", "null"] }
        images:
          type: array
          items: { type: string }
        isFeatured: { type: boolean }
        isAvailable: { type: boolean }
    ProductListResponse:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Product" }
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name: { type: string }
        size: { type: integer }
        contentType: { type: string }
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL: { type: string }
        objectPath: { type: string }