# shared/api-zod/src/generated/types/uploadUrlResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Al Qattan Curtains showroom API
 * OpenAPI spec version: 0.1.0
 */

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}